feat: clamping logic

+++ manenv/vector.py
@@ -2,8 +2,8 @@ import numpy as np
 
 type Vector = type[np.ndarray] 
 
-def make_vector(x : int | float, y : int | float) -> Vector: 
-    return np.array([x, y])
+def make_vector(x : int | float, y : int | float, dtype = int) -> Vector: 
+    return np.array([x, y], dtype=dtype)
 
 class VectorBuiltin:
     ZERO_VECTOR = make_vector(0, 0)
